protecao de rotas

+++ InventarioIFSP/InventarioIFSP/Controllers/ItemStatusController.cs
@@ -17,12 +17,44 @@ namespace InventarioIFSP.Controllers
 
         public ActionResult Create()
         {
+            // Início: Proteção de rota
+            if (Session["usuario_nivel"] == null)
+            {
+                TempData["msg"] = "Necessário estar logado!";
+                TempData["msg_type"] = "warning";
+                return RedirectToAction("Index", "Home");
+            }
+
+            if (Convert.ToInt32(Session["usuario_nivel"]) > 1)
+            {
+                TempData["msg"] = "Você não tem autorização para acessar esta área!";
+                TempData["msg_type"] = "danger";
+                return RedirectToAction("Index", "Home");
+            }
+            // Fim: Proteção de Rota
+
             return View();
         }
 
         [HttpPost]
         public ActionResult Create(ItemStatus cat)
         {
+            // Início: Proteção de rota
+            if (Session["usuario_nivel"] == null)
+            {
+                TempData["msg"] = "Necessário estar logado!";
+                TempData["msg_type"] = "warning";
+                return RedirectToAction("Index", "Home");
+            }
+
+            if (Convert.ToInt32(Session["usuario_nivel"]) > 1)
+            {
+                TempData["msg"] = "Você não tem autorização para acessar esta área!";
+                TempData["msg_type"] = "danger";
+                return RedirectToAction("Index", "Home");
+            }
+            // Fim: Proteção de Rota
+
             try
             {
                 if (ItemStatusDAO.Create(cat) != null)
@@ -44,6 +76,22 @@ namespace InventarioIFSP.Controllers
 
         public ActionResult Edit(int id)
         {
+            // Início: Proteção de rota
+            if (Session["usuario_nivel"] == null)
+            {
+                TempData["msg"] = "Necessário estar logado!";
+                TempData["msg_type"] = "warning";
+                return RedirectToAction("Index", "Home");
+            }
+
+            if (Convert.ToInt32(Session["usuario_nivel"]) > 1)
+            {
+                TempData["msg"] = "Você não tem autorização para acessar esta área!";
+                TempData["msg_type"] = "danger";
+                return RedirectToAction("Index", "Home");
+            }
+            // Fim: Proteção de Rota
+
             try
             {
                 ItemStatus cat = ItemStatusDAO.GetByID(id);
@@ -61,6 +109,22 @@ namespace InventarioIFSP.Controllers
         [HttpPost]
         public ActionResult Edit(ItemStatus status)
         {
+            // Início: Proteção de rota
+            if (Session["usuario_nivel"] == null)
+            {
+                TempData["msg"] = "Necessário estar logado!";
+                TempData["msg_type"] = "warning";
+                return RedirectToAction("Index", "Home");
+            }
+
+            if (Convert.ToInt32(Session["usuario_nivel"]) > 1)
+            {
+                TempData["msg"] = "Você não tem autorização para acessar esta área!";
+                TempData["msg_type"] = "danger";
+                return RedirectToAction("Index", "Home");
+            }
+            // Fim: Proteção de Rota
+
             try
             {
                 if (ItemStatusDAO.Update(status))
@@ -81,6 +145,22 @@ namespace InventarioIFSP.Controllers
         [HttpGet]
         public ActionResult Delete(int id)
         {
+            // Início: Proteção de rota
+            if (Session["usuario_nivel"] == null)
+            {
+                TempData["msg"] = "Necessário estar logado!";
+                TempData["msg_type"] = "warning";
+                return RedirectToAction("Index", "Home");
+            }
+
+            if (Convert.ToInt32(Session["usuario_nivel"]) > 1)
+            {
+                TempData["msg"] = "Você não tem autorização para acessar esta área!";
+                TempData["msg_type"] = "danger";
+                return RedirectToAction("Index", "Home");
+            }
+            // Fim: Proteção de Rota
+
             try
             {
                 if (ItemStatusDAO.Delete(id))
@@ -101,6 +181,22 @@ namespace InventarioIFSP.Controllers
         [HttpGet]
         public ActionResult List()
         {
+            // Início: Proteção de rota
+            if (Session["usuario_nivel"] == null)
+            {
+                TempData["msg"] = "Necessário estar logado!";
+                TempData["msg_type"] = "warning";
+                return RedirectToAction("Index", "Home");
+            }
+
+            if (Convert.ToInt32(Session["usuario_nivel"]) > 1)
+            {
+                TempData["msg"] = "Você não tem autorização para acessar esta área!";
+                TempData["msg_type"] = "danger";
+                return RedirectToAction("Index", "Home");
+            }
+            // Fim: Proteção de Rota
+
             try
             {
                 return View(ItemStatusDAO.GetAll());
